to print theh size of varios types

+++ 0x00-hello_world/6-size.c
@@ -13,11 +13,11 @@ int main(void)
 	long long int ll;
 	float f;
 
-	printf("Size of a char: %llu byte(s)\n", (unsigned long)sizeof(c));
-	printf("Size of an int: %llu byte(s)\n", (unsigned long)sizeof(i));
-	printf("Size of a long int: %llu byte(s)\n", (unsigned long)sizeof(l));
-	printf("Size of a long long int: %llu byte(s)\n", (unsigned long)sizeof(ll));
-	printf("Size of a float: %llu byte(s)\n", (unsigned long)sizeof(f));
+	printf("Size of a char: %d byte(s)\n", sizeof(c));
+	printf("Size of an int: %d byte(s)\n", sizeof(i));
+	printf("Size of a long int: %d byte(s)\n", sizeof(l));
+	printf("Size of a long long int: %d byte(s)\n", sizeof(ll));
+	printf("Size of a float: %d byte(s)\n", sizeof(f));
 
 	return (0);
 }
